refactor(clawdstrike): extract aggregate/bridge/types leaves from engine.rs
Carve the least-coupled leaves of engine.rs into submodules:
- aggregate.rs: severity_to_core + aggregate_overall (thin passthrough to
  the formally-verified core::aggregate; aggregate_overall widened to
  pub(crate) for the in-crate test).
- bridge.rs: BridgeCheckResult + check_bridge_policy + format_origin_brief
  (cross-origin bridge helpers, kept module-internal via private use).
- types.rs: public report types (GuardReport, GuardEvaluationMetadata,
  GuardResolvedEnclave, PostureAwareReport) + EngineStats, re-exported from
  mod.rs so crate::engine::* paths and lib.rs re-exports are unchanged.

Pure relocation: no logic changed; 825 lib tests + formal-diff-tests pass.
